Add host project for Windows 10 IoT Core.
In order to support Windows 10 IoT Core (UWP), I had to retarget all library projects to .NET Standard. Now we have two entry-points: ServiceHost (net472) and IotHost (UWP).

Main issue was the configuration - XamlReader requires main/UI/STA thread to parse/load and there is no access to it from background IoT app. This forced me to roll out my own implementation for XML configuration. I tried to keep it similar to XAML, so that converting existing config is easy.

I also converted virtual file system implementations to async. It probably adds a little bit of the overhead, but since it is I/O - it should be async. This, in turn, required all command handlers to become async, which was a straight-forward, but broad change.